requirements.txt
PyNaCl
click
pydantic
questionary